entity-resolution/requirements-runtime.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
recordlinkage>=0.15
scikit-learn>=1.3.0
jellyfish>=1.0.0
python-Levenshtein>=0.21.0
click>=8.1.0
rich>=13.0.0
